fixed Makefile:
When the deps target was added in November 2020, it became the default
make target. When make was called without arguments, it would install
stuffbin and exit without building hopp-cli.

Added an 'all' target as the default target that installs stuffbin and
builds hopp-cli.

Cleaned up all of the targets.

Changed the deps target to use the idiomatic 'go install' instead of
'go get' to prevent mangling go.mod when installing stuffbin.

Removed the 'update' target as it's redundant re: git pull.